norminette check ok, compilation ok, segfault and memory leak: ok i guess
