+++ includes/cub3D.h
@@ -6,7 +6,7 @@
 /*   By: roaraujo <roaraujo@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/20 11:52:42 by gusalves          #+#    #+#             */
-/*   Updated: 2022/09/27 11:00:43 by roaraujo         ###   ########.fr       */
+/*   Updated: 2022/09/27 12:25:49 by roaraujo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,7 +38,7 @@ typedef struct s_mlx_struct
 	void	*window;
 }				t_mlx_struct;
 
-typedef struct s_map_parameters 
+typedef struct s_map_parameters
 {
 	int		fds[4];
 	int		colours[2][3];
@@ -48,8 +48,13 @@ typedef struct s_map_parameters
 
 // ------------------------------------------------		ENUMS		-----------
 
-enum e_colours {CEILING, FLOOR};
-enum e_directions {NO, SO, EA, WE};
+enum e_colours {ceiling, floor};
+enum e_directions {
+	NO,
+	SO,
+	EA,
+	WE
+};
 enum e_err_codes {
 	SYSCALL_ERROR,
 	WRONG_ARGS_NO,
@@ -67,8 +72,14 @@ void	print_err_exit(int errcode, t_map_parameters *map_params);
 // init_data.c
 void	init_map_parameters(t_map_parameters *map_parameters);
 
+// input_validation_utils.c
+char	*get_next_line_trimmed(int input_fd);
+int		jump_spaces(const char *str);
+int		convert_colour_to_int(char *colour, t_map_parameters *map_params);
+
 // input_validation.c
-void	validate_args(int argc, char *argv[], int *input_fd, t_map_parameters *map_params);
+void	validate_args(int argc, char *argv[], int *input_fd,
+			t_map_parameters *map_params);
 void	validate_input_file(int fd, t_map_parameters *map_params);
 
 // window.c
@@ -80,11 +91,18 @@ int		destroy(t_mlx_struct *mlx);
 // key_press.c
 int		keystrokes_management(int keycode, t_mlx_struct *mlx);
 
+// map_params_checks.c
+void	check_for_valid_path(t_map_parameters *map_params);
+void	check_duplicate_parameter(t_map_parameters *map_params);
+void	check_colour_params_count(char **colours_array,
+			t_map_parameters *map_params);
+void	check_for_valid_colour(t_map_parameters *map_params);
+
 // params_utils.c
 bool	is_direction_identifier(char identifier[2]);
 int		get_direction_index(char d);
 int		get_colour_index(char c);
-bool	is_valid_parameter_char(char  c);
-bool has_valid_param_identifier(char *str);
+bool	is_valid_parameter_char(char c);
+bool	has_valid_param_identifier(char *str);
 
 #endif
